docs: explain external sync and calendar flow
Capture the current Intervals import and calendar data path so the sync model is easier to reason about while keeping the import and polling code split into smaller, reviewable modules.
