+++ src/config/provider_polling/tests/mod.rs
@@ -1,0 +1,6 @@
+mod calendar;
+mod completed_workouts;
+mod persistence;
+mod support;
+
+pub(super) use super::ProviderPollingService;
